fix testComposite - was searching parent inputs
after fixing parent behaviors in Butler the test using
verificationButler was finding a precurosor component because
it was in a parent repo.

I created the verificationButler using a totally separate repo
and wrote the composite with the input-only component into
that separate repo so the input-only object could be successfully
not found.